feat: Add Strava integration
Support fetching the user's Strava token with the introduction of a
StravaContainer. The way this component works is that, until a user is
authenticated against Strava, they'll just be shown a button to log in
Strava. Once they're logged in, any provided children (the rest of the
tree) will be rendered. The children can access the user's Strava token
using the `useStrava()` hook.

This also cleans up the source tree to make things a little more
maintainable and make the API boundaries more clear.

Finally, it also corrects some errors (`for` needs to be `htmlFor`;
`class` needs to be `className`) in the JSX, too.